+++ website/content/about.php
@@ -1,11 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <?php require "../modules/handler.php" ?>
-    <title>Main page - Abscissa</title>
-    <meta name="description" content="Abscissa main page.">
-    <meta name="keywords" content="abscissa main">
+    <?php require_once "../modules/handler.php" ?>
+    <?php add("analytics") ?>
     <?php add("settings") ?>
+    <title>About</title>
+    <meta name="description" content="Abscissa about page.">
+    <meta name="keywords" content="abscissa about">
     <!-- STYLE -->
     <link rel="stylesheet" href="../style/general.css">
     <link rel="stylesheet" href="../style/about.css">