+++ website/modules/analytics.php
@@ -1,0 +1,3 @@
+<!-- ANALYTICS -->
+<script async src="https://www.googletagmanager.com/gtag/js?id=G-LH00FG46KE"></script>
+<script src="../modules/js/analytics.js"></script>