+++ website/modules/environment.php
@@ -1,24 +1,39 @@
 <?php
+// HANDLER
+require_once "../modules/handler.php";
+
+// IMPORTS
+add("exceptions");
+
 function load($request) {
+    // SETS DATA
     $path = __DIR__ . '/../.env';
     $variables = [];
-    if (file_exists($path)) {
-        $content = file($path, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
-        foreach ($content as $line) {
-            // CONTINUE IF COMMENT
-            if (strpos(trim($line), '#') === 0) {
-                continue;
-            }
-            // GET DATA PAIR AND TRIM IT
-            list($key, $value) = explode('=', $line, 2);
-            $key = trim($key);
-            $value = trim($value);
-            // ADD KEY AND VALUE IF NOT EMPTY
-            if (in_array($key, $request)) {
-                $variables[$key] = $value;
-            }
+    // CHECK IF FILE EXISTS
+    if (!file_exists($path)) {
+        throw new environmentNotFoundException("");
+    }
+    // GETS CONTENT
+    $content = file($path, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+    // ITERATES OVER EACH LINE
+    foreach ($content as $line) {
+        // CONTINUE IF COMMENT
+        if (strpos(trim($line), '#') === 0) {
+            continue;
+        }
+        // GET DATA PAIR AND TRIM IT
+        list($key, $value) = explode('=', $line, 2);
+        $key = trim($key);
+        $value = trim($value);
+        // STORE VARIABLE
+        $variables[$key] = $value;
+    }
+    // CHECK IF REQUESTED VARIABLES ARE PRESENT
+    foreach ($request as $variable) {
+        if (!array_key_exists($variable, $variables)) {
+            throw new environmentVariableNotFoundException("$variable");
         }
     }
-    return $variables;
+    return array_intersect_key($variables, array_flip($request));
 }
 ?>