+++ website/modules/handler.php
@@ -1,6 +1,11 @@
 <?php
+// REQUIRE A MODULE FUNCTION
 function add($module) {
-    $path = __DIR__ . "/../modules/$module.php";
-    include $path;
+    try {
+        $path = __DIR__ . "/../modules/$module.php";
+        require_once $path;
+    } catch (Exception $error) {
+        echo "Failed to import $module.";
+    }
 }
 ?>