+++ website/modules/js/analytics.js
@@ -1,0 +1,5 @@
+// GOOGLE ANALYTICS
+window.dataLayer = window.dataLayer || [];
+function gtag(){dataLayer.push(arguments);}
+gtag('js', new Date());
+gtag('config', 'G-LH00FG46KE');