+++ website/script/general.js
@@ -1,4 +1,0 @@
-// REDIRECT FUNCTION
-function redirect(target) {
-    window.location.href = `./${target}.php`
-}